Add support for metric views to enable downsampling features
This change gives users the ability to register metric views which
will represent a new form of metric data in the system as a new metric.
Views have to be derived from a metric that is already in the system
and cannot be ingested to. Querying the metric view data can be done
only by specifying the exact metric view name in the label matcher.
This is a first step into adding a robust downsampling solution for
Promscale, later steps will remove some of the restrictions and give
more freedom in how the metric view data is represented.